requirements.txt
sqlparse
pydantic
pydantic-settings
gitpython
requests
chromadb
langchain
langchain-community
tiktoken
sentence-transformers
langchain-openai
langchain-google-genai
fastapi
uvicorn
python-multipart
python-dotenv
tenacity
deep-translator
sqlalchemy
psycopg2-binary
pymysql
